fix: remove PartialEq and Eq from Vector trait

+++ crates/quantization/src/quantize.rs
@@ -369,7 +369,7 @@ mod reduce_sum_of_x_as_u32 {
 #[inline(always)]
 pub fn quantize(lut: &[f32], n: f32) -> (f32, f32, Vec<u8>) {
     let (min, max) = f32::reduce_min_max_of_x(lut);
-    let k = 0.0f32.max((max - min) / (n as f32));
+    let k = 0.0f32.max((max - min) / n);
     let b = min;
     (k, b, mul_add_round::mul_add_round(lut, 1.0 / k, -b / k))
 }
